v2.2
Add config file
Add time for file name
Add print file name
Rename some modules
Update README
