+++ README.md
@@ -84,39 +84,26 @@ pip install re
 <a name="run"></a>
 <h2>Виклик та завантаження</h2>
 
-Щоб розпізнати мову з відео необхідно на підготовчому етапі помістити відповідні відеофайли `.avi` у директорію `/videos`/.
-
-Для серверів, які працюють під керівництвом операційних систем сімейства `Windows OS`, виклик програмної системи **"language-recognizer"** здійснюється шляхом запуску скрипта `ім'я скрипта.py` з використанням команди `python`. Потрібно відкрити командний рядок – термінал `cmd` та написати `python ім'я скрипта.py`. Важливо, щоб скрипт знаходився або в директорії, з якої запущено командний рядок, або в каталозі, прописаному у змінній середовища `PATH`. 
-Тож завантаження програмної системи забезпечується введенням в командному рядку повного імені завантажувальної програми без додаткових параметрів (шлях до директорії з відео- чи аудіофайлами, мову мовлення яких потрібно розпізнати, задано за замовчуванням `/videos/` та `/audios/`, відповідно):
-```sh
-python main.py
-```
-
-Для серверів, які працюють під керівництвом `Unix`-подібних операційних систем (наприклад, `Linux`) також можна скористатися цим способом, але на початку скрипта `Python` у першому рядку має бути вказаний повний шлях до інтерпретатора:
-```sh
-#!/usr/bin/python3
-```
-або
-```sh
-#!/usr/bin/env python3
-```
-
-Для роботи з програмною системою необхідно клонувати репозиторій та проінсталювати залежності:
+Для роботи з програмною системою необхідно:
+1. Використовуючи консоль, завдяки наступним командам клонувати репозиторій та перейти у робочу дерикторію:
 ```sh
 git clone https://github.com/OlehDmytrenko/language-recognizer.git
 cd language-recognizer
 ```
-Запуск програмної системи виконується за допомогою введення в терміналі ім'я головного виконуваного скрипта ```main.py```:
+або вивантажити цей репозиторій [`language-recognizer`](https://github.com/OlehDmytrenko/language-recognizer/archive/refs/heads/main.zip) вручну, розпакувати його, зупустити консоль та перейти у корінь робочої директорії:
+```sh
+cd language-recognizer
+```
+2. Далі, щоб ідентифікувати мову з відео, необхідно на підготовчому етапі помістити відповідні відеофайли `.avi` у директорію `./videos`, що повинна знаходитись у корені робочої директорії `language-recognizer` (створити директорію `./videos` у випадку її відсутності).
+3. Виклик програмної системи без додаткових параметрів забезпечується введенням в терміналі ім'я головного виконуваного скрипта ```main.py``` (шлях до директорії з відео- чи аудіофайлами, мову мовлення яких потрібно розпізнати, задано за замовчуванням - `/videos/`):
 ```sh
 python main.py
 ```
-
-Запуск програмної системи з зовнішніми налаштуваннями виконується шляхом введення цих налаштувань після імені головного виконуваного скрипта `main.py` з додатковими параметрами `[<path to settings>]`:
+Виклик програмної системи з зовнішніми налаштуваннями виконується шляхом введення цих налаштувань після імені головного виконуваного скрипта `main.py` з додатковими параметрами `[<path to settings>]`:
 ```sh
 python <path to script> [<path to settings>]
 
 ```
-
 В якості вхідного параметра може бути повний шлях до відео- /...повний шлях.../імʼя_відеофайла.mp4 чи аудіофайла `/....повний шлях.../імʼя_аудіофайла.wav/`. В такому випадку, завантаження програмної системи забезпечується введенням в командному рядку повного імені головного виконуваного скрипта `main.py` з додатковими параметрами:
 ```sh
 python main.py ./...повний шлях.../імʼя_відеофайла.mp4
@@ -126,16 +113,16 @@ python main.py ./...повний шлях.../імʼя_відеофайла.mp4
 python main.py ./...повний шлях.../імʼя_аудіофайла.wav
 ```
 
-В результаті запуску скрипта `language-recognizer.py` здійснюється попередня конвертація вхідних відеофайлів формату `.avi` у аудіо формату `.wav`. Вихідні аудіофайли буде автоматично збережено у директорії `/audios`/ та додано до черги для ідентифікації представленої у них мови мовлення.
-Далі програмна система по черзі вибиратиме файли для ідентифікації мови із директорії `/audios/` та виводитиме результат у консоль.
+В результаті запуску скрипта `main.py` здійснюється попередня конвертація вхідних відеофайлів формату `.avi` у аудіо формату `.wav`, розбиття їх на менші частини з вказаною тривалістю (за замовчуванням формуються аудіофайли тривалістю по `15` секунд). Вихідні аудіофайли буде автоматично збережено у директорії `/audios` та додано до черги для ідентифікації представленої у них мови мовлення.
+Далі програмна система по черзі вибиратиме файли для ідентифікації мови із директорії `/audios/` та виводитиме результат у вигляді назви розпізнаного аудіофайла `<audioName>.wav` та мітки `Ukrainian`, `Russian` чи `Mixed` у консоль.
 
-Дані, отримані в результаті застосування програмної системи та які стосуються основних результатів виводяться в консолі. 
-Також вивід може перенаправлятися із консолі у файл, який зберігаюється у директорії `results` у вигляді `.log` файла. Для цього використовується оператор `>`.
+Дані, які стосуються проміжних результатів, отриманих в процесі виконання програмної системи, виводяться у файл `outlog.log`. 
+
+Також вивід може перенаправлятися із консолі у файл, який зберігаюється у кореневій директорії у вигляді `.log` файла. Для цього використовується оператор `>`.
 Повна команда виглядає так:
 ```sh
 python main.py > outlog.log
 ```
-Тут `outlog.log` – це текстовий файл, у який записуються проміжні результати та  виконання програмної системи.
 
 Операція може використовуватися як в операційній системі `Windows OS`, так і в `Unix`-подібних системах.
 Якщо файла, в який повинен вивестися результат, не існує – система створить його автоматично.
@@ -143,5 +130,7 @@ python main.py > outlog.log
 ```sh
 python language-recognizer.py >> outlog.log
 ```
+
+Під час першого запуску програми, у випадку відсутності української чи російської моделей ([`vosk-model-uk-v3`](https://alphacephei.com/vosk/models/vosk-model-uk-v3.zip) та [`vosk-model-small-ru-0.22`(https://alphacephei.com/vosk/models/vosk-model-small-ru-0.22.zip), відповідно), що повинні знаходитись у директорії `./models`, система автоматично здійснить завантаження та розпакування цих моделей. Цей процес ресурсозадратний й може вимагати достатньо багато часу на його виконання.
 
 © 2022 [Oleh Dmytrenko](https://github.com/OlehDmytrenko)
