+++ config.json
@@ -1,0 +1,8 @@
+{"langModels" : [
+	{"uk" : "vosk-model-uk-v3"},
+	{"ru" : "vosk-model-ru-0.22"}
+	],
+
+"audioDuration" : "30"
+
+}
