--- defaultLangs.csv
@@ -1,2 +1,0 @@
-uk
-ru
