--- defaultModelsHashMap.csv
@@ -1,2 +1,0 @@
-uk:vosk-model-uk-v3
-ru:vosk-model-ru-0.22
